modal_flows/pyproject.toml
[tool.poetry]
name = "modal_flows"
version = "0.1.0"
description = ""
authors = ["feliche93 <felix.vemmer@gmail.com>"]
readme = "README.md"
packages = [
    { include = "app" },
    { include = "papierkram" },
]

[tool.poetry.dependencies]
python = ">=3.12.2,<=3.13"
playwright = "^1.40.0"
python-dotenv = "^1.0.0"
clean-text = {extras = ["gpl"], version = "^0.6.0"}
langdetect = "^1.0.9"
pandas = "^2.1.3"
beautifulsoup4 = "^4.12.2"
gender-guesser = "^0.4.0"
unstructured = "^0.11.0"
tiktoken = "^0.5.1"
googlesearch-python = "^1.2.3"
fastapi = "^0.104.1"
markdownify = "^0.11.6"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
html2text = "^2020.1.16"
customerio = "^2.1"
python-telegram-bot = "^20.6"
playwright-stealth = "^1.0.6"
psycopg2-binary = "^2.9.9"
dateparser = "^1.2.0"
modal = "^0.56.4376"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
marvin = "^2.3.0"
sqlmodel = "^0.0.16"
pendulum = "^3.0.0"
tenacity = "^8.2.3"
langchain = "^0.1.13"
langsmith = "^0.1.31"
instructor = "^1.0.0"
anthropic = "^0.21.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.6"
ipykernel = "^6.26.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
